uint64_t -> uint32_t

+++ include/Core/Constants.h
@@ -42,21 +42,21 @@ inline const std::string EX_SZ_A = "SZ";
  */
 namespace OrderType {
 // 限价单，指定价格的挂单，订单时限为当日有效
-inline const uint64_t LIMIT = 1;
+inline const uint32_t LIMIT = 1;
 
 // 市价单，不同的平台对市价单的支持不同
-inline const uint64_t MARKET = 2;
+inline const uint32_t MARKET = 2;
 
 // 对方最优价格的限价单
-inline const uint64_t BEST = 3;
+inline const uint32_t BEST = 3;
 
 // 立即全部成交否则立即撤销，不同平台的实现也不同
 // 对于CTP来说是限价单，需要指定价格
-inline const uint64_t FAK = 4;
+inline const uint32_t FAK = 4;
 
 // 立即成交剩余部分立即撤单，不同平台的实现也不同
 // 对于CTP来说是限价单，需要指定价格
-inline const uint64_t FOK = 5;
+inline const uint32_t FOK = 5;
 }  // namespace OrderType
 
 /*
@@ -64,16 +64,16 @@ inline const uint64_t FOK = 5;
  */
 namespace Direction {
 // 买入
-inline const uint64_t BUY = 1;
+inline const uint32_t BUY = 1;
 
 // 卖出
-inline const uint64_t SELL = 2;
+inline const uint32_t SELL = 2;
 
 // 申购，暂时不支持
-inline const uint64_t PURCHASE = 4;
+inline const uint32_t PURCHASE = 4;
 
 // 赎回，暂时不支持
-inline const uint64_t REDEEM = 5;
+inline const uint32_t REDEEM = 5;
 }  // namespace Direction
 
 /*
@@ -83,34 +83,34 @@ inline const uint64_t REDEEM = 5;
  */
 namespace Offset {
 // 开仓
-inline const uint64_t OPEN = 1;
+inline const uint32_t OPEN = 1;
 
 // 平仓
-inline const uint64_t CLOSE = 2;
+inline const uint32_t CLOSE = 2;
 
 // 平今
-inline const uint64_t CLOSE_TODAY = 4;
+inline const uint32_t CLOSE_TODAY = 4;
 
 // 平昨
-inline const uint64_t CLOSE_YESTERDAY = 8;
+inline const uint32_t CLOSE_YESTERDAY = 8;
 }  // namespace Offset
 
 /*
  * 对手方，只针买卖方向有效
  */
-inline uint64_t opp_direction(uint64_t d) {
+inline uint32_t opp_direction(uint32_t d) {
   return d == Direction::BUY ? Direction::SELL : Direction::BUY;
 }
 
 /*
  * 是否是开仓
  */
-inline bool is_offset_open(uint64_t offset) { return offset == Offset::OPEN; }
+inline bool is_offset_open(uint32_t offset) { return offset == Offset::OPEN; }
 
 /*
  * 是否是平仓，平仓包括平仓、平今、平昨
  */
-inline bool is_offset_close(uint64_t offset) {
+inline bool is_offset_close(uint32_t offset) {
   return offset &
          (Offset::CLOSE | Offset::CLOSE_TODAY | Offset::CLOSE_YESTERDAY);
 }
@@ -118,8 +118,8 @@ inline bool is_offset_close(uint64_t offset) {
 /*
  * 交易方向转为string
  */
-inline const std::string& direction_str(uint64_t d) {
-  static const std::map<uint64_t, std::string> d_str = {
+inline const std::string& direction_str(uint32_t d) {
+  static const std::map<uint32_t, std::string> d_str = {
       {Direction::BUY, "Buy"}, {Direction::SELL, "Sell"}};
 
   return d_str.find(d)->second;
@@ -128,8 +128,8 @@ inline const std::string& direction_str(uint64_t d) {
 /*
  * 开平类型转为string
  */
-inline const std::string& offset_str(uint64_t off) {
-  static const std::map<uint64_t, std::string> off_str = {
+inline const std::string& offset_str(uint32_t off) {
+  static const std::map<uint32_t, std::string> off_str = {
       {Offset::OPEN, "Open"},
       {Offset::CLOSE, "Close"},
       {Offset::CLOSE_TODAY, "CloseToday"},
@@ -141,8 +141,8 @@ inline const std::string& offset_str(uint64_t off) {
 /*
  * 订单价格类型转string
  */
-inline const std::string& ordertype_str(uint64_t t) {
-  static const std::map<uint64_t, std::string> t_str = {
+inline const std::string& ordertype_str(uint32_t t) {
+  static const std::map<uint32_t, std::string> t_str = {
       {OrderType::LIMIT, "Limit"},
       {OrderType::MARKET, "Market"},
       {OrderType::BEST, "Best"},
