+++ include/Core/TradingEngineInterface.h
@@ -48,13 +48,13 @@ class TradingEngineInterface {
   /*
    * 订单成交时回调
    */
-  virtual void on_order_traded(uint64_t order_id, int64_t this_traded,
+  virtual void on_order_traded(uint64_t order_id, int this_traded,
                                double traded_price) {}
 
   /*
    * 撤单成功时回调
    */
-  virtual void on_order_canceled(uint64_t order_id, int64_t canceled_volume) {}
+  virtual void on_order_canceled(uint64_t order_id, int canceled_volume) {}
 
   /*
    * 撤单被拒时回调
